requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
importlib-metadata==0.17
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pytest==4.6.2
pytest-django==3.5.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.5.1
